+++ src/Network.jl
@@ -50,7 +50,7 @@ end
 
 function read_boundary(path::String)
     # Read Fortran formatted extent file. It determines the limit on the nuclear chart.
-    raw_boundary::Matrix{Int64} = readdlm(path,' ',Int)
+    raw_boundary::Matrix{Int64} = readdlm(path, ' ', Int64)
     fill_boundary(raw_boundary)
 end
